+++ NEWS.md
@@ -1,5 +1,18 @@
 # Release notes for cmsaf
 
+## 3.2.0
+
+- Add new operators timavg, num_above, num_below, num_equal,
+  mon_num_above, mon_num_below, mon_num_equal
+- Add operators cmsaf.mul and cmsaf.div to Toolbox
+- New slider in visualizer to change image ratio
+- Adaptation to source information for figures
+
+## 3.1.1
+
+- Minor update of dates and mail address
+- Bug fix in box_mergetime and get_time due to unit 'seconds since'
+
 ## 3.1.0
 
 - Extend functionalities of Climate Analysis